+++ CHANGES.rst
@@ -1,6 +1,12 @@
 ChangeLog
 =========
 
+1.2
+---
+* **FEATURE** Method to validate a specific component in keyMap
+* Helper method to reconstruct a schema endpoint from the path to root
+* Additional documentation and unittest improvements
+
 1.1
 ---
 * Bug fix for index out of range error created from empty list input
@@ -13,8 +19,6 @@ ChangeLog
 TODO
 ----
 * Update loader to use join instead of home brew path method
-* Method to create a blank instance of model
-* Method to validate a specific component in keyMap (ignore references to other keys)
 * Allow declaration of multiple datatypes for key values in maps
 * Validate size of data object inputs
 * Allow declaration of a lambda function for further validation